modelo_recomendador/requirements.txt
ydata-profiling==4.8.3
seaborn==0.13.2
scikit-learn==1.5.0
pandas==2.2.2
mlflow==2.14.1
joblib==1.4.2


isort
keras
jupyter
black
autoflake
evidently
snowflake-connector-python
mlxtend
fastapi
uvicorn
pytest
orjson